Replace `int` indices with `size_t`
Loops of the form (i = max-1; i >= 0; i--) were replaced using the
equivalent form (i = max; i > 0; i--) where uses of `i` were replaced with
`i-1`. This is the minimally invasive replacement which avoids the common
mistake of looping with the condition `i >= 0` which is always true when
using an unsigned loop variable.